Refactored user event emitting.
Specific fields from the data object are now iterated through and
checked for changes. The event names, etc. are derived from the field
names. This gives us a bit less code to maintain.

Also the events now include the actor who caused the event, ie. who
moved the user to a channel or who muted them.