Version 0.1.0
Some backwards compatibility breaking changes in client.sendMessage. We
had two `sendMessage` implementations in the APIs:

- `MumbleConnection.sendMessage` and `MumbleClient.sendMessage`

  These two methods were used to send raw protobuf messages to the
  server. The `MumbleClient.sendMessage` delegated directly to the
  former.
- `User.sendMessage` and `Channel.sendMessage`

  These two methods were used to send text messages to users/channels.

Since the `MumbleClient` is closer to `User` and `Channel` in high level
API usage it made more sense to have the `sendMessage` on the client to
behave identical to the ones on `User` and `Channel`.

The old method can still be found from `client.connection.sendMessage`.