Added channelByPath() to MumbleClient
client.channelByPath( path ) can be used to query a signle channel by
its path. The path uses '/' as segment separator and the channel names
are URL encoded. Essentially this should be compatible with Mumble's
"Get URL" command.

The path can also be given unencoded as long as it doesn't contain
escape codes that would be mangled by the unencode process.