utility function for rescaling a frame
The MumbleInputStream requires frames to be of 16 Bit deep signed values. With this function 8 Bit deep unsigned or signed value frames can be upscaled to 16 Bit signed. It should also work with 16 Bit unsigned to 16 Bit signed and downscaling from 32 Bit, but that's not tested (and probably unneccessary as well).